requirements.txt
pyTelegramBotAPI==4.6.1
python-dotenv==0.20.0
pyzabbix==1.1.0
